Remove config_client fixture as no longer needed
