+++ system_tests/conftest.py
@@ -1,5 +1,2 @@
-# Add run_engine and mock_config_client to be used in system tests
-pytest_plugins = [
-    "dodal.testing.fixtures.run_engine",
-    "dodal.testing.fixtures.config_client",
-]
+# Add run_engine to be used in system tests
+pytest_plugins = ["dodal.testing.fixtures.run_engine"]
